Add social filter block

+++ database/migrations/2019_06_29_104345_create_scripts_table.php
@@ -16,7 +16,7 @@ class CreateScriptsTable extends Migration
         Schema::create('scripts', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
 
             $table->string('title');
