+++ database/migrations/2019_06_30_091414_create_blocks_table.php
@@ -16,7 +16,7 @@ class CreateBlocksTable extends Migration
         Schema::create('blocks', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('schema_id');
+            $table->unsignedBigInteger('schema_id');
 
             $table->morphs('data');
