+++ database/migrations/2019_07_02_124021_create_receive_message_blocks_table.php
@@ -18,7 +18,7 @@ class CreateReceiveMessageBlocksTable extends Migration
 
             $table->string('message')->nullable();
 
-            $table->integer('next_block_id')->nullable();
+            $table->unsignedBigInteger('next_block_id')->nullable();
 
             $table->timestamps();
