+++ database/migrations/2019_07_04_092515_create_send_message_blocks_table.php
@@ -18,7 +18,7 @@ class CreateSendMessageBlocksTable extends Migration
 
             $table->string('message')->nullable();
 
-            $table->integer('next_block_id')->nullable();
+            $table->unsignedBigInteger('next_block_id')->nullable();
 
             $table->timestamps();
