+++ database/migrations/2019_07_04_092516_create_send_message_with_keyboard_blocks_table.php
@@ -18,7 +18,7 @@ class CreateSendMessageWithKeyboardBlocksTable extends Migration
 
             $table->string('message')->nullable();
 
-            $table->integer('error_next_block_id')->nullable();
+            $table->unsignedBigInteger('error_next_block_id')->nullable();
 
             $table->timestamps();
