+++ database/migrations/2019_07_16_090934_create_reset_param_blocks_table.php
@@ -18,9 +18,9 @@ class CreateResetParamBlocksTable extends Migration
 
             $table->boolean('reset_all')->default(false);
 
-            $table->integer('param_id')->nullable();
+            $table->unsignedBigInteger('param_id')->nullable();
 
-            $table->integer('next_block_id')->nullable();
+            $table->unsignedBigInteger('next_block_id')->nullable();
 
             $table->timestamps();
